msm: camera: isp: Modify RDI only state machine for Waipio
In Waipio, the sequence of the RUP and SOF interrupts is reversed as
compared to previous generations. RUP is observed before SOF.

This requires handling of RUP in applied state, which is missing in the
RDI only state machine leading to the state machine looping to the
bubble state repeatedly.

By adding the transition from applied to epoch state on RUP, looping
over the bubble state is no longer observed.

CRs-Fixed: 2846451
Change-Id: Ib298695ae2aad248b76a6e05c6a9b4d84793717b
Signed-off-by: Anand Ravi <ananravi@codeaurora.org>