msm: camera: isp: Add fixes for vfe and sfe stride programming
Currently, image stride programming in line based mode programs
the unaligned value from io config directly to config register.
If the value coming from user space is not aligned, buf done will
not be seen for that WM. Throw a warning in this case.
Also, LTM Stats port is programmed to be in frame based mode,
change it to line based mode to reflect correct configuration.

CRs-Fixed: 2854696
Change-Id: I71b4e0576dfda7dfe025dbdb81bdb9462b82704a
Signed-off-by: Mukund Madhusudan Atre <matre@codeaurora.org>