msm: camera: isp: Reduce IRQ top half IO operations
In the IRQ controller top half, registers which do not belong to any
event should not be read since they are not going to be processed
further.

Also, issue the global clear after top half processing to accommodate
cases where two controllers share a common global clear and one is
triggered by the other. This change adds support for this by introducing
functions to explicitly define such dependencies.

CRs-Fixed: 2846451
Change-Id: Ie5df596bc3f342550c08f172647ad0cb5ed3f3c7
Signed-off-by: Anand Ravi <ananravi@codeaurora.org>