msm: camera: isp: Enable IFE Bus violation IRQ
This commit corrects the bit mask for enabling the Bus violation
IRQ. Bus side errors were not caught early as the bit mask was not
correct.
As a part of this commit, we cause the hw to stop when bus side
violation errors are reported. This prevents the false reporting
of CSID back pressure overflows.

Change-Id: I1770ff14e186f252da9b871e0c29fe5f4c9fab0a
CRs-Fixed: 2997739
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>