msm: camera: isp: Read bus consumed addr at bus irq top half
Problem Description:

Due to scheduling delay, We observed a scenario where we have a request
N in active list, and N+1 is also applied on the output frame for N.

It is possible that after applying N+1th, the buf done tasklet for N is
not scheduled in time, and if it so happens that the tasklet is
scheduled out beyond the next frame, HW would have consumed N+1, and we
end up reading the last consumed addr for N+1 in the buf done bh for N.

The read last consumed address from N+1, will never match with N
ultimately stalling N.

Solution:

We can avoid this scenario if we read the last consumed address
registers in top half.

Move the read of consumed addr to bus top half, as the tasklet
may delay and cannot readout the consumed addr in time.

CRs-Fixed: 3342249
Change-Id: I9742b117f433df6275ae5058d62e77ffce1f9a3f
Signed-off-by: chengxue <quic_chengxue@quicinc.com>
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>