bus: mhi: update buffer type to u8
Normally, arithmetic operations on void pointers are not allowed in C,
as the compiler needs to know the exact size of the object the pointer
void * is pointing to. GCC extension allows it to be used as an
exception. If Someone decides to use a different compiler which does not
support the arithmetic operations on void pointers, the corresponding
implementation may fail the compile.

In order to reduce the likelihood of bugs being introduced in the code
and improve code quality. This change updates the buffer
pointer type from void * to u8 * wherever applicable, and the corresponding
implementation.

Change-Id: Id0760f68239f60ca9ff8d766ee0b5c4694c47a6c
Signed-off-by: Arnav Kumar Dawn <quic_arnavkum@quicinc.com>