kgsl: hwsched: Add support for GMU tracepoints logging
In hardware scheduling some events originating in GMU
are not captured on host side. The preemption tracepoints
for preempt trigger and preempt done are examples. Such
tracepoints need to be continued to be supported with
HW scheduling for backwards compatibility and profiling
or debugging. We therefore need a means to be able to log
these events on the GMU with the timestamp of when they
occur and convey them to the host such that kgsl can log
them lazily to ftrace.

Change-Id: Ib12e2341f928091ad3918841c267a8f2e92dc766
Signed-off-by: Hareesh Gundu <quic_hareeshg@quicinc.com>
Signed-off-by: NISARG SHETH <nsheth@qti.qualcomm.com>