sched/walt: Add support for multiple top-app cgroups
Currently, only tasks under the "top-app" cgroup (/dev/cpuctl/top-app)
are considered for colocation, along with other boost settings.
Add support for considering "normal" and "latency-sensitive"
cgroups as part of the top-app group. These cgroups are required to
be present at /dev/cpuctl/apps/top-apps/normal and
/dev/cpuctl/apps/top-apps/latency-sensitive, respectively.
This enhancement allows tasks within these subgroups to benefit from the
privileges of the top-app group.
Note, a task in any of the following cgroups will be considered as
top-app:
1. /dev/cpuctl/top-app/tasks
2. /dev/cpuctl/apps/top-app/tasks
3. /dev/cpuctl/apps/top-app/normal/tasks
4. /dev/cpuctl/apps/top-app/latency-sensitive/tasks.

Change-Id: Ifcd3a84a2ddd26e8f83e2bf75735a8d0ad7dec6e
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>